[feat]: 푸터 하나로 통일
- Landing 푸터 지움
- 포인터 추가
