+++ src/pages/landing/LandingPage.tsx
@@ -7,8 +7,8 @@ import {
   LandingEmotion,
   LandingAnswer,
   LandingCatchPhrase,
-  LandingFooter,
-  LandingHeader
+  LandingHeader,
+  Footer
 } from 'components'
 import styled from 'styled-components'
 import { motion } from 'framer-motion'
@@ -25,7 +25,7 @@ const LandingPage = () => {
           <LandingEmotion />
           <LandingAnswer />
           <LandingCatchPhrase />
-          <LandingFooter />
+          <Footer />
         </LandingInfo>
       </LandingWrapper>
     </>
